jvs_auth_92kx-main/Jarvis- Aula 01/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
ddgs
requests
python-dotenv
pycaw
comtypes
screen_brightness_control
playwright
yt-dlp
pygetwindow
pyautogui
numpy
sounddevice
# GUI Automation (Novas dependências)
pywinauto
pillow
keyboard
pyperclip